Improves DX for field renaming.
At the moment, field renaming involves exporting configuration files in the
*live* site after running the relevant database updates.  This is different
from the usual deployment process where database updates are run in a
development machine and then *drush deploy* or equivalent is used by a CI/CD
system to deploy Drupal.

This is happening because every time the database updates for field renaming are
run, they generate different UUID value for field configs which instantly makes
the new field config out-of-sync with any config file on the disk.  To avoid
this, we should source the UUIDs from config files on the disk *when* such files
are present.  This makes the process *drush deploy* compatible.