With the view "files" disabled, localgov_core cause an exception #283  (#297)
* With the view "files" disabled, localgov_core cause an exception

* With the view "files" disabled, localgov_core cause an exception - comment empty catch

* The intention of #221 was to move the files tab from the main content menu.

Note: This is opinionated administration configuration that probably
shouldn't be enforced on users just requiring core or media. It probably
wants to live somewhere else.

* Link to the files view under media. As per #221. See also #283 issues.

The view is supplied by localgov_media so creating the link here too.
Probably wants to be in some other LocalGov Admin configuration module
really.

* Coding standards fix

---------

Co-authored-by: Marco Pagliarulo <marcopagliarulo@gmail.com>
Co-authored-by: Stephen Cox <stephen@agile.coop>